+++ docs/changes.rst
@@ -7,6 +7,8 @@ Version 0.3.2
 To be released.
 
 - Fix incorrect encoding of filenames.  [:issue:`122`]
+- Fix key type of :attr:`Image.metadata <wand.image.Image.metadata>`
+  dictionary to :class:`str` from :class:`bytes` in Python 3.
 
 
 Version 0.3.1